+++ changelog.md
@@ -23,6 +23,14 @@
 
 ### Breaking Changes
 
+- in `Data.Morpheus.Server`:
+
+  - removed `subscriptionApp`
+  - changed `webSocketsApp` type to `App e m -> m (ServerApp, e -> m ())`
+  - changed `httpPubApp` type to `[e -> m ()] -> App e m -> a -> m b`
+
+- removed `Stream` from `Data.Morpheus.Types`
+
 - removed class `Interpreter`, `interpreter` is now just regular function.
 
   ```hs